+++ ui/src/pages/admin/settings/org.tsx
@@ -154,16 +154,10 @@ class EditOrg extends React.Component<Props, State> {
     if (this.state.saved) {
       hint = <Alert variant="success">{this.props.t("entryUpdated")}</Alert>;
     } else if (this.state.error) {
-      hint = (
-        <Alert variant="danger">
-          {this.props.t("errorSave")}
-          <br />
-          {this.props.t("hintErrorSaveOrg")}
-        </Alert>
-      );
+      hint = <Alert variant="danger">{this.props.t("errorSave")}</Alert>;
     }
 
-    let buttonSave = (
+    const buttonSave = (
       <Button
         className="btn-sm"
         variant="outline-secondary"